Move backend tests to backend/open_webui/test/ per spec
- Move unit tests: routers/test_models.py, models/test_models_table.py,
  utils/test_models_cache.py → backend/open_webui/test/
- Move integration tests: test_cache_router.py, test_access_control.py,
  test_super_admin_assign.py → backend/open_webui/test/apps/webui/routers/
- Merge unit + integration conftest into single backend/open_webui/test/conftest.py
- Update workflow: paths trigger on backend/**, pytest targets backend/open_webui/test/
- Update OpenShift runner to glob *.live.spec.ts (exclude mocked specs)
- Update tests/README.md with new paths and structure
