+++ backend/open_webui/test/apps/webui/models/test_models_table.py
