+++ backend/open_webui/test/conftest.py
@@ -1,17 +1,13 @@
-"""Shared fixtures for backend integration tests.
+"""Shared fixtures for custom model backend tests (unit + integration).
 
-Uses SQLite (same pattern as unit backend conftest). Sets env vars before
-any open_webui imports to avoid triggering real Postgres or Redis connections.
+Env vars must be set BEFORE any open_webui imports that trigger database
+initialization (open_webui.config.py runs migrations at import time).
 
-ENV VARS (override via shell or CI — do NOT hardcode in production):
-  WEBUI_SECRET_KEY                       — required, fail-fast if missing in CI
-  SUPER_ADMIN_EMAILS                     — required, fail-fast if missing in CI
-  USER_PERMISSIONS_WORKSPACE_MODELS_ACCESS — optional, defaults to "true"
-  DATABASE_URL                           — optional, defaults to local SQLite
-
-NOTE: This conftest is intentionally separate from the unit conftest.
-      Different DATABASE_URL and SUPER_ADMIN_EMAILS prevent combining
-      unit + integration in a single pytest invocation.
+ENV VARS (override via shell or CI):
+  WEBUI_SECRET_KEY                        — defaults to test value
+  SUPER_ADMIN_EMAILS                      — defaults to test value
+  USER_PERMISSIONS_WORKSPACE_MODELS_ACCESS — defaults to "true"
+  DATABASE_URL                            — defaults to local SQLite
 """
 
 import os
@@ -22,7 +18,6 @@ _CI = os.environ.get("CI", "").lower() in ("1", "true", "yes")
 
 
 def _require_env(name: str, default: str | None = None) -> str:
-    """Return env var value. Fail fast in CI if missing and no default."""
     val = os.environ.get(name)
     if val:
         return val
@@ -34,17 +29,11 @@ def _require_env(name: str, default: str | None = None) -> str:
     return ""
 
 
-os.environ["WEBUI_SECRET_KEY"] = _require_env("WEBUI_SECRET_KEY", "test-secret-key-for-integration-tests")
+os.environ["WEBUI_SECRET_KEY"] = _require_env("WEBUI_SECRET_KEY", "test-secret-key-for-tests")
 os.environ["SUPER_ADMIN_EMAILS"] = _require_env("SUPER_ADMIN_EMAILS", "admin@example.com")
-
-# Allow non-admin users to create models (workspace permission)
 os.environ.setdefault("USER_PERMISSIONS_WORKSPACE_MODELS_ACCESS", "true")
-# Do NOT override REDIS_URL — let it use the default redis://localhost:6379/0
-# CacheManager gracefully handles connection failures via _check_redis_available()
 
-_TEST_DB_PATH = os.path.join(
-    os.path.dirname(__file__), "..", "..", "..", "test_integration.db"
-)
+_TEST_DB_PATH = os.path.join(os.path.dirname(__file__), "test_custom_models.db")
 _TEST_DB_URI = f"sqlite:///{_TEST_DB_PATH}"
 open(_TEST_DB_PATH, "a").close()
 os.environ.setdefault("DATABASE_URL", _TEST_DB_URI)
@@ -52,12 +41,12 @@ os.environ.setdefault("DATABASE_URL", _TEST_DB_URI)
 import pytest
 from contextlib import contextmanager
 
-# Unique prefix for this test run — all integration test data should use this
 TEST_RUN_PREFIX = f"test-custom-models-{int(time.time())}"
 
 
 def pytest_configure(config):
     config.option.asyncio_mode = "auto"
+    config.addinivalue_line("markers", "unit: unit tests (backend)")
     config.addinivalue_line("markers", "integration: integration tests (backend)")
 
 
@@ -113,6 +102,34 @@ def seed_admin(db_session):
         role="admin",
     )
     yield user
+
+
+@pytest.fixture
+def seed_user(db_session):
+    from open_webui.models.users import Users
+    user = Users.insert_new_user(
+        id="test-user-1",
+        name="Test User",
+        email="test@example.com",
+        profile_image_url="/user.png",
+        role="admin",
+    )
+    yield user
+    Users.delete_user_by_id("test-user-1")
+
+
+@pytest.fixture
+def seed_user_2(db_session):
+    from open_webui.models.users import Users
+    user = Users.insert_new_user(
+        id="test-user-2",
+        name="Test User 2",
+        email="test2@example.com",
+        profile_image_url="/user.png",
+        role="user",
+    )
+    yield user
+    Users.delete_user_by_id("test-user-2")
 
 
 @pytest.fixture
@@ -191,6 +208,40 @@ def mock_admin_user(app):
 
 
 @pytest.fixture
+def mock_regular_user(app):
+    @contextmanager
+    def _mock(user_id="user-1", email="user@test.com", role="user"):
+        from open_webui.utils.auth import (
+            get_verified_user,
+            get_current_user,
+            get_current_user_by_api_key,
+        )
+        from open_webui.models.users import User
+
+        def create_user():
+            return User(
+                id=user_id,
+                name="Regular User",
+                email=email,
+                role=role,
+                profile_image_url="/user.png",
+                last_active_at=0,
+                updated_at=0,
+                created_at=0,
+            )
+
+        app.dependency_overrides = {
+            get_current_user: create_user,
+            get_verified_user: create_user,
+            get_current_user_by_api_key: create_user,
+        }
+        yield
+        app.dependency_overrides = {}
+
+    return _mock
+
+
+@pytest.fixture
 def mock_user_a(app):
     @contextmanager
     def _mock(user_id="user-a", email="a@example.com", role="user"):
@@ -260,14 +311,6 @@ def mock_user_b(app):
 
 @pytest.fixture
 def cleanup_groups():
-    """Track and clean up groups created during a test.
-
-    Usage:
-        def test_something(cleanup_groups):
-            group_id = cleanup_groups("my-group-id")
-            ... test logic ...
-            # group auto-deleted after test (even on failure)
-    """
     created: list[str] = []
 
     def _register(group_id: str) -> str:
@@ -286,13 +329,6 @@ def cleanup_groups():
 
 @pytest.fixture
 def cleanup_functions():
-    """Track and clean up functions created during a test.
-
-    Usage:
-        def test_something(cleanup_functions):
-            func_id = cleanup_functions("my-func-id")
-            ... test logic ...
-    """
     created: list[str] = []
 
     def _register(func_id: str) -> str:
@@ -311,7 +347,6 @@ def cleanup_functions():
 
 @pytest.fixture
 def unique_model_id():
-    """Generate a unique model ID for integration tests using the run prefix."""
     def _make(suffix: str = "") -> str:
         parts = [TEST_RUN_PREFIX]
         if suffix:
