+++ scripts/run_openshift_frontend_quality_checks.sh
@@ -74,7 +74,7 @@ forward_metrics_to_grafana_cloud() {
 playwright_status=0
 npx playwright test \
   playwright/tests/ai-tutor-dashboard.live.spec.ts \
-  playwright/tests/models/ \
+  playwright/tests/models/*.live.spec.ts \
   --workers="${PLAYWRIGHT_WORKERS}" \
   --retries="${PLAYWRIGHT_RETRIES}" \
   --timeout="${PLAYWRIGHT_TIMEOUT}" || playwright_status=$?
